requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
ultralytics
opencv-python-headless
pillow
python-multipart
